Remove `else None` sugar for basic cases of detectable exhaustive match.
Construct a union type of all cases in the match that match on type alone,
and if this union is a supertype of the match expression type,
then the match is exhaustive so we don't need to sugar the `else None`.

As a result, users won't have to deal with the type system possibility of a `None`
result for match blocks which can easily be shown to never reach the `else` case.